backend/requirements.txt
torch
opencv-python
pillow
numpy
git+https://github.com/openai/CLIP.git


